+++ Firmware/GMT-Geiger-Counter/CosmicRayDetector.h
@@ -28,7 +28,7 @@ class CosmicRayDetector {
 
     bool              _enabled;             // Flag for checking if the cosmic ray detector is enabled
     volatile uint16_t _movingAverage[60];   // Array for storing coincidence events per hour
-    volatile uint16_t _movingAverageIndex;  // Index of the moving average array
+    volatile uint8_t  _movingAverageIndex;  // Index of the moving average array
     hw_timer_t        *_movingAverageTimer; // Hardware timer for advancing the moving average array
     Tube              _coincidenceTube;     // Virtual coincidence tube
