+++ Firmware/GMT-Geiger-Counter/Tube.h
@@ -12,7 +12,7 @@ class Tube {
   public:
 
     // Constructor
-    Tube(const uint8_t pin, volatile uint16_t *movingAverage, volatile uint16_t &movingAverageIndex);
+    Tube(const uint8_t pin, volatile uint16_t *movingAverage, volatile uint8_t &movingAverageIndex);
 
     void begin();            // Initialize everything
     void enable();           // Enable the tube
@@ -26,7 +26,7 @@ class Tube {
 
     const uint8_t     _pin;                    // The pin, the tube is connected to
     volatile uint16_t *_movingAverage;         // Pointer to the moving average array
-    volatile uint16_t &_movingAverageIndex;    // Reference to the moving average array index
+    volatile uint8_t  &_movingAverageIndex;    // Reference to the moving average array index
     bool              _enabled;                // Flag for checking if tube is enabled
     volatile uint64_t _pulseTimerMicroseconds; // Timer for measuring the pulse length
     volatile uint64_t _counts;                 // Total number of counts
